chore:增加会话API
